Rename openexr_conf.h to openexr_config.h
